# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "end": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End"
          },
          "initial_capital": {
            "default": 10000.0,
            "title": "Initial Capital",
            "type": "number"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "start": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start"
          },
          "strategy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuyAndHoldConfig"
              },
              {
                "$ref": "#/components/schemas/PriceVsMAConfig"
              }
            ],
            "title": "Strategy"
          }
        },
        "required": [
          "instrument_id",
          "strategy"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "BacktestRequest": {
        "properties": {
          "data_source": {
            "default": "yfinance",
            "enum": [
              "yfinance",
              "vnstock",
              "csv"
            ],
            "title": "Data Source",
            "type": "string"
          },
          "date_range": {
            "$ref": "#/components/schemas/DateRange"
          },
          "initial_capital": {
            "default": 10000.0,
            "title": "Initial Capital",
            "type": "number"
          },
          "max_leverage": {
            "default": 1.0,
            "title": "Max Leverage",
            "type": "number"
          },
          "strategy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BuyAndHoldConfig"
              },
              {
                "$ref": "#/components/schemas/PriceVsMAConfig"
              }
            ],
            "title": "Strategy"
          },
          "symbols": {
            "items": {
              "type": "string"
            },
            "title": "Symbols",
            "type": "array"
          }
        },
        "required": [
          "symbols",
          "date_range",
          "strategy"
        ],
        "title": "BacktestRequest",
        "type": "object"
      },
      "BuyAndHoldConfig": {
        "properties": {
          "type": {
            "const": "buy_and_hold",
            "default": "buy_and_hold",
            "title": "Type",
            "type": "string"
          },
          "weight": {
            "default": 1.0,
            "title": "Weight",
            "type": "number"
          }
        },
        "title": "BuyAndHoldConfig",
        "type": "object"
      },
...
      "CompanyCatalogResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyCatalogItemResponse"
            },
            "title": "Companies",
            "type": "array"
          },
          "facets": {
            "$ref": "#/components/schemas/CompanyCatalogFacetsResponse"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "offset",
          "limit",
          "companies",
          "facets"
        ],
        "title": "CompanyCatalogResponse",
        "type": "object"
      },
...
      "CompanyListResponse": {
        "properties": {
          "as_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "As Of"
          },
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyResponse"
            },
            "title": "Companies",
            "type": "array"
          },
          "country_code": {
            "enum": [
              "US",
              "VN"
            ],
            "title": "Country Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "facets": {
            "$ref": "#/components/schemas/CompanyListFacetsResponse"
          },
          "fetched_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fetched At"
          },
          "id": {
            "enum": [
              "US_ALL",
              "US100",
              "US2000",
              "US500",
              "US30",
              "VN_ALL",
              "VNALL",
              "VN100",
              "VN30",
              "VNMID",
              "VNSML"
            ],
            "title": "Id",
            "type": "string"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "name",
          "country_code",
          "description",
          "total",
          "offset",
          "limit",
          "companies",
          "facets"
        ],
        "title": "CompanyListResponse",
        "type": "object"
      },
...
      "DateRange": {
        "properties": {
          "end": {
            "format": "date",
            "title": "End",
            "type": "string"
          },
          "start": {
            "format": "date",
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end"
        ],
        "title": "DateRange",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PortfolioResultResponse": {
        "properties": {
          "equity_curve": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Equity Curve",
            "type": "object"
          },
          "final_nav": {
            "title": "Final Nav",
            "type": "number"
          },
          "total_return_pct": {
            "title": "Total Return Pct",
            "type": "number"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeSchema"
            },
            "title": "Trades",
            "type": "array"
          },
          "weights": {
            "additionalProperties": {
              "additionalProperties": {
                "type": "number"
              },
              "type": "object"
            },
            "title": "Weights",
            "type": "object"
          }
        },
        "required": [
          "equity_curve",
          "trades",
          "weights",
          "total_return_pct",
          "final_nav"
        ],
        "title": "PortfolioResultResponse",
        "type": "object"
      },
...
      "PriceVsMAConfig": {
        "description": "Price vs Moving Average strategy.\n\nGoes long when close > MA, flat otherwise.\nbuy_lag / sell_lag: confirmation bars required before entry / exit.",
        "properties": {
          "buy_lag": {
            "default": 0,
            "title": "Buy Lag",
            "type": "integer"
          },
          "ma_length": {
            "default": 50,
            "title": "Ma Length",
            "type": "integer"
          },
          "ma_type": {
            "default": "sma",
            "enum": [
              "sma",
              "ema",
              "wma"
            ],
            "title": "Ma Type",
            "type": "string"
          },
          "sell_lag": {
            "default": 0,
            "title": "Sell Lag",
            "type": "integer"
          },
          "type": {
            "const": "price_vs_ma",
            "default": "price_vs_ma",
            "title": "Type",
            "type": "string"
          }
        },
        "title": "PriceVsMAConfig",
        "type": "object"
      },
...
      "SingleTickerAnalysisResponse": {
        "properties": {
          "bah": {
            "$ref": "#/components/schemas/PerformanceSummaryResponse"
          },
          "current_position": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CurrentPositionResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_last_session": {
            "format": "date",
            "title": "Data Last Session",
            "type": "string"
          },
          "equity_curve_bah": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Equity Curve Bah",
            "type": "object"
          },
          "equity_curve_strategy": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Equity Curve Strategy",
            "type": "object"
          },
          "expected_last_session": {
            "format": "date",
            "title": "Expected Last Session",
            "type": "string"
          },
          "from_date": {
            "title": "From Date",
            "type": "string"
          },
          "health_by_year": {
            "items": {
              "$ref": "#/components/schemas/HealthRowResponse"
            },
            "title": "Health By Year",
            "type": "array"
          },
          "instrument_id": {
            "title": "Instrument Id",
            "type": "integer"
          },
          "is_stale": {
            "title": "Is Stale",
            "type": "boolean"
          },
          "mae_percentiles_winners": {
            "items": {
              "$ref": "#/components/schemas/DistributionRowResponse"
            },
            "title": "Mae Percentiles Winners",
            "type": "array"
          },
          "mfe_percentiles_losers": {
            "items": {
              "$ref": "#/components/schemas/DistributionRowResponse"
            },
            "title": "Mfe Percentiles Losers",
            "type": "array"
          },
          "mfe_percentiles_winners": {
            "items": {
              "$ref": "#/components/schemas/DistributionRowResponse"
            },
            "title": "Mfe Percentiles Winners",
            "type": "array"
          },
          "monthly_returns_bah": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Monthly Returns Bah",
            "type": "object"
          },
          "monthly_returns_strategy": {
            "additionalProperties": {
              "additionalProperties": {
                "anyOf": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "type": "object"
            },
            "title": "Monthly Returns Strategy",
            "type": "object"
          },
          "monthly_stats_by_calendar": {
            "items": {
              "$ref": "#/components/schemas/MonthlyStatRowResponse"
            },
            "title": "Monthly Stats By Calendar",
            "type": "array"
          },
          "monthly_stats_by_entry_month": {
            "items": {
              "$ref": "#/components/schemas/MonthlyStatRowResponse"
            },
            "title": "Monthly Stats By Entry Month",
            "type": "array"
          },
          "price_basis": {
            "title": "Price Basis",
            "type": "string"
          },
          "price_source": {
            "title": "Price Source",
            "type": "string"
          },
          "refresh_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refresh Warning"
          },
          "refreshed": {
            "title": "Refreshed",
            "type": "boolean"
          },
          "return_percentiles": {
            "items": {
              "$ref": "#/components/schemas/DistributionRowResponse"
            },
            "title": "Return Percentiles",
            "type": "array"
          },
          "strategy": {
            "$ref": "#/components/schemas/PerformanceSummaryResponse"
          },
          "strategy_label": {
            "title": "Strategy Label",
            "type": "string"
          },
          "symbol": {
            "title": "Symbol",
            "type": "string"
          },
          "ticker_prices": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Ticker Prices",
            "type": "object"
          },
          "to_date": {
            "title": "To Date",
            "type": "string"
          },
          "total_bars": {
            "title": "Total Bars",
            "type": "integer"
          },
          "trades": {
            "items": {
              "$ref": "#/components/schemas/TradeRowResponse"
            },
            "title": "Trades",
            "type": "array"
          },
          "undercut_distribution": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/UndercutDistributionRowResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undercut Distribution"
          },
          "venue_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Venue Code"
          }
        },
        "required": [
          "symbol",
          "strategy_label",
          "from_date",
          "to_date",
          "total_bars",
          "current_position",
          "strategy",
          "bah",
          "trades",
          "return_percentiles",
          "mae_percentiles_winners",
          "mfe_percentiles_winners",
          "mfe_percentiles_losers",
          "monthly_returns_strategy",
          "monthly_returns_bah",
          "monthly_stats_by_calendar",
          "monthly_stats_by_entry_month",
          "health_by_year",
          "equity_curve_strategy",
          "equity_curve_bah",
          "ticker_prices",
          "instrument_id",
          "expected_last_session",
          "data_last_session",
          "refreshed",
          "is_stale",
          "price_source",
          "price_basis"
        ],
        "title": "SingleTickerAnalysisResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Backtesting and factor analysis API",
    "title": "Trading Statistics Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/backtest": {
      "post": {
        "operationId": "run_backtest_backtest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortfolioResultResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Backtest",
        "tags": [
          "backtest"
        ]
      }
    },
    "/backtest/analyze": {
      "post": {
        "description": "Full analytics for a single-ticker strategy: performance, trades, heatmaps, health.",
        "operationId": "analyze_single_ticker_backtest_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleTickerAnalysisResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Single Ticker",
        "tags": [
          "backtest"
        ]
      }
    },
    "/companies": {
      "get": {
        "operationId": "listCompanies",
        "parameters": [
          {
            "in": "query",
            "name": "universe",
            "required": false,
            "schema": {
              "default": "US_ALL",
              "enum": [
                "US_ALL",
                "US100",
                "US2000",
                "US500",
                "US30",
                "VN_ALL",
                "VNALL",
                "VN100",
                "VN30",
                "VNMID",
                "VNSML"
              ],
              "title": "Universe",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "in": "query",
            "name": "industry",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          },
          {
            "in": "query",
            "name": "venue",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Venue"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5000,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Companies",
        "tags": [
          "companies"
        ]
      }
    },
    "/companies/catalog": {
      "get": {
        "operationId": "listCompanyCatalog",
        "parameters": [
          {
            "in": "query",
            "name": "country",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "US",
                    "VN"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sector",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 255,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 5000,
              "maximum": 5000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyCatalogResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Company Catalog",
        "tags": [
          "companies"
        ]
      }
    },
...
  }
}